Update captureImage.py
The SURF and SIFT algorithms are actually patented. The ORB object is not, so in order to avoid problems, we will use that. ORB is also considered a fast alternative to SURF/SIFT algorithms. So we will use the ORB object to find key points and descriptors, and the flann based matcher to find matches.